refactor: simplify file structure to flat organization
Consolidate nested module structure into 5 focused files:
- Merge layout/* into single bsp.rs (BSP tree + algorithms)
- Merge config/* into single config.rs (config + validation)
- Merge keyboard.rs + keys.rs into keyboard.rs (key handling)
- Merge window_manager/* into single window_manager.rs (window management)
- Update main.rs to remove lib.rs dependency

Benefits:
- Reduced from 15 files to 5 files
- Clearer code organization with section comments
- All 59 tests passing, zero clippy warnings
- All functionality preserved
- Easier navigation and maintenance
