--- src/lib.rs
@@ -1,10 +1,0 @@
-//! Rustile - A tiling window manager written in Rust
-//!
-//! This window manager provides automatic window tiling with a master-stack layout.
-//! It's designed to be simple, efficient, and extensible.
-
-pub mod config;
-pub mod keyboard;
-pub mod keys;
-pub mod layout;
-pub mod window_manager;
